+++ source/frontend/components/pages/Overview.vue
@@ -11,6 +11,7 @@
       <f7-list-item title="Collection" link="/collection/"></f7-list-item>
       <f7-list-item title="MySQL API" link="/mysqlapi/"></f7-list-item>
       <f7-list-item title="MySQL Collection" link="/mysqlCollection/"></f7-list-item>
+      <f7-list-item title="Google Auth" link="/googleAuth/"></f7-list-item>
     </f7-list>
   </f7-page>
 </template>