+++ CHANGELOG.md
@@ -7,6 +7,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.39] - 2022-12-29
+
+### Added
+
+- onGoogleAuth
+
+### Changed
+
+- useLocalStorage has null as default value
+- useLocalStorage will remove items on null value
+
+### Removed
+
+- onGoogleSignin
+
 ## [v1.38.1] - 2022-12-29
 
 ### Fixed